Revert "Revert "merge master""
This reverts commit 4748e4c05b9f31cf2bcd5ec7e47eb37487c168fa.

+++ i386/i386/idt.c
@@ -30,7 +30,7 @@ struct real_gate idt[IDTSZ];
 
 struct idt_init_entry
 {
-	unsigned entrypoint;
+	unsigned long entrypoint;
 	unsigned short vector;
 	unsigned short type;
 };
